pyproject.toml
[build-system]
requires = ["uv_build>=0.10.9,<0.11"]
build-backend = "uv_build"

[project]
name = "leafpress"
version = "0.8.1"
description = "Convert MkDocs sites to PDF, Word, HTML, ODT, EPUB, and Markdown documents with branding"
readme = "README.md"
authors = [{name = "Shane Hutchins", email = "hutchins@users.noreply.github.com"}]
maintainers = [
  {name = "Shane Hutchins", email = "hutchins@users.noreply.github.com"}
]
requires-python = ">=3.13"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Text Processing :: Markup :: Markdown",
]
keywords = ["mkdocs", "pdf", "docx", "documentation", "converter"]

dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "markdown>=3.5",
    "pymdown-extensions>=10.0",
    "pygments>=2.17",
    "python-docx>=1.1.0",
    "gitpython>=3.1.40",
    "jinja2>=3.1",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "requests>=2.32",
    "python-dotenv>=1.0",
    "odfpy>=1.4.1",
    "ebooklib>=0.18",
    "mammoth>=1.8",
    "markdownify>=1.0",
    "python-pptx>=1.0.0",
    "openpyxl>=3.1",
    "pylatexenc>=2.10",
]

[project.optional-dependencies]
pdf = [
    "weasyprint>=62.0",
]
ui = [
    "pyqt6>=6.6.0",
    "pyobjc-framework-cocoa>=10.0; sys_platform == 'darwin'",
]
all = [
    "leafpress[pdf,ui]",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
    "mypy>=1.10",
    "weasyprint>=62.0",
]

[project.scripts]
leafpress = "leafpress.cli:cli"

[project.urls]
Homepage = "https://leafpress.dev"
Repository = "https://github.com/hutchins/leafpress"

[tool.uv]
package = true

[tool.uv.build-backend]
module-root = "src"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
"src/leafpress/cli.py" = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["tests"]
addopts = "-v --tb=short"
markers = ["docker: tests requiring Docker (may be slow)"]

[tool.coverage.run]
source = ["leafpress"]
omit = ["src/leafpress/ui/*"]

[tool.coverage.report]
show_missing = true
skip_empty = true
fail_under = 70

[tool.ty.environment]
extra-paths = ["tests"]

[[tool.ty.overrides]]
include = ["src/leafpress/ui/", "src/leafpress/doctor.py", "src/leafpress/package_version.py"]
[tool.ty.overrides.rules]
unresolved-import = "ignore"

[tool.mypy]
python_version = "3.10"
strict = true